Refactor DataAccess layer and fix security vulnerability
This commit addresses two major issues in the DataAccess layer:

1.  **Fix SQL Injection Vulnerability:**
    The `GetPreguntasSeguridadByIds` method used unsafe string concatenation to build an SQL query, making it vulnerable to SQL injection. This has been refactored to use a parameterized query, eliminating the security risk.

2.  **Refactor Repositories to Adhere to SRP:**
    The `SqlUserRepository` and its interface violated the Single Responsibility Principle by managing nearly every entity in the application. This "God Repository" has been broken down into smaller, more focused repositories:
    *   `IUserRepository`: Now only manages `Usuario` entities.
    *   `IPersonaRepository`: Manages `Persona` entities.
    *   `ISecurityRepository`: Manages security-related entities like policies and questions.
    *   `IReferenceDataRepository`: Manages read-only lookup data.

    The corresponding services, factories, and dependency injection configurations in the `BusinessLogic` layer have been updated to use these new, more specific repository interfaces. All existing unit tests were updated and are passing, ensuring the refactoring did not break existing functionality.
