+++ src/DataAccess/Repositories/SqlPersonaRepository.cs
@@ -1,0 +1,156 @@
+using System;
+using System.Collections.Generic;
+using System.Data;
+using DataAccess.Entities;
+using Microsoft.Data.SqlClient;
+using Microsoft.Extensions.Logging;
+using UserManagementSystem.DataAccess.Exceptions;
+
+namespace DataAccess.Repositories
+{
+    public class SqlPersonaRepository : IPersonaRepository
+    {
+        private readonly DatabaseConnectionFactory _connectionFactory;
+        private readonly ILogger<SqlPersonaRepository> _logger;
+
+        public SqlPersonaRepository(DatabaseConnectionFactory connectionFactory, ILogger<SqlPersonaRepository> logger)
+        {
+            _connectionFactory = connectionFactory ?? throw new ArgumentNullException(nameof(connectionFactory));
+            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+        }
+
+        private T ExecuteReader<T>(string sql, Func<SqlDataReader, T> map, Action<SqlParameterCollection>? addParameters = null, CommandType commandType = CommandType.Text)
+        {
+            using (var connection = (SqlConnection)_connectionFactory.CreateConnection())
+            {
+                connection.Open();
+                using (var command = connection.CreateCommand())
+                {
+                    command.CommandText = sql;
+                    command.CommandType = commandType;
+                    addParameters?.Invoke(command.Parameters);
+
+                    using (var reader = command.ExecuteReader())
+                    {
+                        return map(reader);
+                    }
+                }
+            }
+        }
+
+        private void ExecuteNonQuery(string sql, Action<SqlParameterCollection> addParameters, CommandType commandType = CommandType.StoredProcedure)
+        {
+            try
+            {
+                using (var connection = (SqlConnection)_connectionFactory.CreateConnection())
+                {
+                    connection.Open();
+                    using (var command = connection.CreateCommand())
+                    {
+                        command.CommandText = sql;
+                        command.CommandType = commandType;
+                        addParameters(command.Parameters);
+                        command.ExecuteNonQuery();
+                    }
+                }
+            }
+            catch (SqlException ex)
+            {
+                _logger.LogError(ex, "Ocurrió un error de SQL al ejecutar ExecuteNonQuery para el comando: {sql}", sql);
+                throw new DataAccessLayerException($"Ocurrió un error de SQL al ejecutar {sql}", ex);
+            }
+        }
+
+        public Persona? GetPersonaById(int id) => ExecuteReader("SELECT * FROM personas WHERE id_persona = @id;", reader =>
+        {
+            if (!reader.Read()) return null;
+            return new Persona(
+                (int)reader["legajo"],
+                reader["nombre"] as string ?? string.Empty,
+                reader["apellido"] as string ?? string.Empty,
+                (int)reader["id_tipo_doc"],
+                reader["num_doc"] as string ?? string.Empty,
+                reader["fecha_nacimiento"] as DateTime?,
+                reader["cuil"] as string,
+                reader["calle"] as string,
+                reader["altura"] as string,
+                (int)reader["id_localidad"],
+                (int)reader["id_genero"],
+                reader["correo"] as string,
+                reader["celular"] as string,
+                (DateTime)reader["fecha_ingreso"]
+            );
+        }, p => p.AddWithValue("@id", id));
+
+        public List<Persona> GetAllPersonas() => ExecuteReader("SELECT * FROM personas;", reader =>
+        {
+            var list = new List<Persona>();
+            while (reader.Read())
+            {
+                list.Add(new Persona(
+                    (int)reader["legajo"],
+                    reader["nombre"] as string ?? string.Empty,
+                    reader["apellido"] as string ?? string.Empty,
+                    (int)reader["id_tipo_doc"],
+                    reader["num_doc"] as string ?? string.Empty,
+                    reader["fecha_nacimiento"] as DateTime?,
+                    reader["cuil"] as string,
+                    reader["calle"] as string,
+                    reader["altura"] as string,
+                    (int)reader["id_localidad"],
+                    (int)reader["id_genero"],
+                    reader["correo"] as string,
+                    reader["celular"] as string,
+                    (DateTime)reader["fecha_ingreso"]
+                ));
+            }
+            return list;
+        });
+
+        public void AddPersona(Persona persona)
+        {
+            _logger.LogInformation("Agregando persona a la base de datos: Legajo={Legajo}", persona.Legajo);
+            ExecuteNonQuery("sp_insert_persona", p =>
+            {
+                p.AddWithValue("@legajo", persona.Legajo);
+                p.AddWithValue("@nombre", persona.Nombre);
+                p.AddWithValue("@apellido", persona.Apellido);
+                p.AddWithValue("@id_tipo_doc", persona.IdTipoDoc);
+                p.AddWithValue("@num_doc", persona.NumDoc);
+                p.AddWithValue("@fecha_nacimiento", (object?)persona.FechaNacimiento ?? DBNull.Value);
+                p.AddWithValue("@cuil", (object?)persona.Cuil ?? DBNull.Value);
+                p.AddWithValue("@calle", (object?)persona.Calle ?? DBNull.Value);
+                p.AddWithValue("@altura", (object?)persona.Altura ?? DBNull.Value);
+                p.AddWithValue("@id_localidad", persona.IdLocalidad);
+                p.AddWithValue("@id_genero", persona.IdGenero);
+                p.AddWithValue("@correo", (object?)persona.Correo ?? DBNull.Value);
+                p.AddWithValue("@celular", (object?)persona.Celular ?? DBNull.Value);
+            });
+            _logger.LogInformation("Persona con legajo {Legajo} agregada exitosamente.", persona.Legajo);
+        }
+
+        public void UpdatePersona(Persona persona) => ExecuteNonQuery("sp_update_persona", p =>
+        {
+            p.AddWithValue("@id_persona", persona.IdPersona);
+            p.AddWithValue("@legajo", persona.Legajo);
+            p.AddWithValue("@nombre", persona.Nombre);
+            p.AddWithValue("@apellido", persona.Apellido);
+            p.AddWithValue("@id_tipo_doc", persona.IdTipoDoc);
+            p.AddWithValue("@num_doc", persona.NumDoc);
+            p.AddWithValue("@fecha_nacimiento", (object?)persona.FechaNacimiento ?? DBNull.Value);
+            p.AddWithValue("@cuil", (object?)persona.Cuil ?? DBNull.Value);
+            p.AddWithValue("@calle", (object?)persona.Calle ?? DBNull.Value);
+            p.AddWithValue("@altura", (object?)persona.Altura ?? DBNull.Value);
+            p.AddWithValue("@id_localidad", persona.IdLocalidad);
+            p.AddWithValue("@id_genero", persona.IdGenero);
+            p.AddWithValue("@correo", (object?)persona.Correo ?? DBNull.Value);
+            p.AddWithValue("@celular", (object?)persona.Celular ?? DBNull.Value);
+            p.AddWithValue("@fecha_ingreso", (object?)persona.FechaIngreso ?? DBNull.Value);
+        });
+
+        public void DeletePersona(int personaId) => ExecuteNonQuery("sp_delete_persona", p =>
+        {
+            p.AddWithValue("@id_persona", personaId);
+        });
+    }
+}
