+++ src/DataAccess/Repositories/SqlUserRepository.cs
@@ -64,168 +64,6 @@ namespace DataAccess.Repositories
             }
         }
 
-        public List<TipoDoc> GetAllTiposDoc() => ExecuteReader("SELECT id_tipo_doc, tipo_doc FROM tipo_doc;", reader =>
-        {
-            var list = new List<TipoDoc>();
-            while (reader.Read())
-            {
-                list.Add(new TipoDoc { IdTipoDoc = (int)reader["id_tipo_doc"], Nombre = (string)reader["tipo_doc"] });
-            }
-            return list;
-        });
-
-        public List<Genero> GetAllGeneros() => ExecuteReader("SELECT id_genero, genero FROM generos;", reader =>
-        {
-            var list = new List<Genero>();
-            while (reader.Read())
-            {
-                list.Add(new Genero { IdGenero = (int)reader["id_genero"], Nombre = (string)reader["genero"] });
-            }
-            return list;
-        });
-
-        public List<Rol> GetAllRoles() => ExecuteReader("SELECT id_rol, rol FROM roles;", reader =>
-        {
-            var list = new List<Rol>();
-            while (reader.Read())
-            {
-                list.Add(new Rol { IdRol = (int)reader["id_rol"], Nombre = (string)reader["rol"] });
-            }
-            return list;
-        });
-
-        public List<Provincia> GetAllProvincias() => ExecuteReader("SELECT id_provincia, provincia FROM provincias;", reader =>
-        {
-            var list = new List<Provincia>();
-            while (reader.Read())
-            {
-                list.Add(new Provincia { IdProvincia = (int)reader["id_provincia"], Nombre = (string)reader["provincia"] });
-            }
-            return list;
-        });
-
-        public List<Partido> GetPartidosByProvinciaId(int provinciaId) => ExecuteReader("SELECT id_partido, partido, id_provincia FROM partidos WHERE id_provincia = @id_provincia;", reader =>
-        {
-            var list = new List<Partido>();
-            while (reader.Read())
-            {
-                list.Add(new Partido { IdPartido = (int)reader["id_partido"], Nombre = (string)reader["partido"], IdProvincia = (int)reader["id_provincia"] });
-            }
-            return list;
-        }, p => p.AddWithValue("@id_provincia", provinciaId));
-
-        public List<Localidad> GetLocalidadesByPartidoId(int partidoId) => ExecuteReader("SELECT id_localidad, localidad, id_partido FROM localidades WHERE id_partido = @id_partido;", reader =>
-        {
-            var list = new List<Localidad>();
-            while (reader.Read())
-            {
-                list.Add(new Localidad { IdLocalidad = (int)reader["id_localidad"], Nombre = (string)reader["localidad"], IdPartido = (int)reader["id_partido"] });
-            }
-            return list;
-        }, p => p.AddWithValue("@id_partido", partidoId));
-
-        public PoliticaSeguridad? GetPoliticaSeguridad() => ExecuteReader("SELECT TOP 1 * FROM politicas_seguridad;", reader =>
-        {
-            if (!reader.Read()) return null;
-            return new PoliticaSeguridad(
-                (int)reader["id_politica"],
-                reader["mayus_y_minus"] as bool? ?? false,
-                reader["letras_y_numeros"] as bool? ?? false,
-                reader["caracter_especial"] as bool? ?? false,
-                reader["autenticacion_2fa"] as bool? ?? false,
-                reader["no_repetir_anteriores"] as bool? ?? false,
-                reader["sin_datos_personales"] as bool? ?? false,
-                reader["min_caracteres"] as int? ?? 8,
-                reader["cant_preguntas"] as int? ?? 3
-            );
-        });
-
-        public List<PreguntaSeguridad> GetPreguntasSeguridad() => ExecuteReader("SELECT id_pregunta, pregunta FROM preguntas_seguridad;", reader =>
-        {
-            var list = new List<PreguntaSeguridad>();
-            while (reader.Read())
-            {
-                list.Add(new PreguntaSeguridad { IdPregunta = (int)reader["id_pregunta"], Pregunta = (string)reader["pregunta"] });
-            }
-            return list;
-        });
-
-        public List<PreguntaSeguridad> GetPreguntasSeguridadByIds(List<int> ids)
-        {
-            if (ids == null || !ids.Any())
-            {
-                return new List<PreguntaSeguridad>();
-            }
-
-            var idList = string.Join(",", ids);
-            var sql = $"SELECT id_pregunta, pregunta FROM preguntas_seguridad WHERE id_pregunta IN ({idList})";
-
-            return ExecuteReader(sql, reader =>
-            {
-                var list = new List<PreguntaSeguridad>();
-                while (reader.Read())
-                {
-                    list.Add(new PreguntaSeguridad { IdPregunta = (int)reader["id_pregunta"], Pregunta = (string)reader["pregunta"] });
-                }
-                return list;
-            });
-        }
-
-        public TipoDoc? GetTipoDocByNombre(string nombre) => ExecuteReader("SELECT id_tipo_doc, tipo_doc FROM tipo_doc WHERE tipo_doc = @nombre;", reader =>
-        {
-            if (!reader.Read()) return null;
-            return new TipoDoc { IdTipoDoc = (int)reader["id_tipo_doc"], Nombre = (string)reader["tipo_doc"] };
-        }, p => p.AddWithValue("@nombre", nombre));
-
-        public Localidad? GetLocalidadByNombre(string nombre) => ExecuteReader("SELECT id_localidad, localidad, id_partido FROM localidades WHERE localidad = @nombre;", reader =>
-        {
-            if (!reader.Read()) return null;
-            return new Localidad { IdLocalidad = (int)reader["id_localidad"], Nombre = (string)reader["localidad"], IdPartido = (int)reader["id_partido"] };
-        }, p => p.AddWithValue("@nombre", nombre));
-
-        public Genero? GetGeneroByNombre(string nombre) => ExecuteReader("SELECT id_genero, genero FROM generos WHERE genero = @nombre;", reader =>
-        {
-            if (!reader.Read()) return null;
-            return new Genero { IdGenero = (int)reader["id_genero"], Nombre = (string)reader["genero"] };
-        }, p => p.AddWithValue("@nombre", nombre));
-
-        public Rol? GetRolByNombre(string nombre) => ExecuteReader("SELECT id_rol, rol FROM roles WHERE rol = @nombre;", reader =>
-        {
-            if (!reader.Read()) return null;
-            return new Rol { IdRol = (int)reader["id_rol"], Nombre = (string)reader["rol"] };
-        }, p => p.AddWithValue("@nombre", nombre));
-
-        public Persona? GetPersonaById(int id) => ExecuteReader("SELECT * FROM personas WHERE id_persona = @id;", reader =>
-        {
-            if (!reader.Read()) return null;
-            return new Persona(
-                (int)reader["legajo"],
-                reader["nombre"] as string ?? string.Empty,
-                reader["apellido"] as string ?? string.Empty,
-                (int)reader["id_tipo_doc"],
-                reader["num_doc"] as string ?? string.Empty,
-                reader["fecha_nacimiento"] as DateTime?,
-                reader["cuil"] as string,
-                reader["calle"] as string,
-                reader["altura"] as string,
-                (int)reader["id_localidad"],
-                (int)reader["id_genero"],
-                reader["correo"] as string,
-                reader["celular"] as string,
-                (DateTime)reader["fecha_ingreso"]
-            );
-        }, p => p.AddWithValue("@id", id));
-
-        public List<RespuestaSeguridad>? GetRespuestasSeguridadByUsuarioId(int idUsuario) => ExecuteReader("SELECT id_usuario, id_pregunta, respuesta FROM respuestas_seguridad WHERE id_usuario = @id_usuario;", reader =>
-        {
-            var list = new List<RespuestaSeguridad>();
-            while (reader.Read())
-            {
-                list.Add(new RespuestaSeguridad { IdUsuario = (int)reader["id_usuario"], IdPregunta = (int)reader["id_pregunta"], Respuesta = (string)reader["respuesta"] });
-            }
-            return list;
-        }, p => p.AddWithValue("@id_usuario", idUsuario));
-
         public List<HistorialContrasena> GetHistorialContrasenasByUsuarioId(int idUsuario) => ExecuteReader("SELECT id, id_usuario, fecha_cambio, contrasena_script FROM historial_contrasena WHERE id_usuario = @id_usuario;", reader =>
         {
             var list = new List<HistorialContrasena>();
@@ -303,64 +141,10 @@ namespace DataAccess.Repositories
             return list;
         }, commandType: CommandType.StoredProcedure);
 
-        public List<Persona> GetAllPersonas() => ExecuteReader("SELECT * FROM personas;", reader =>
-        {
-            var list = new List<Persona>();
-            while (reader.Read())
-            {
-                list.Add(new Persona(
-                    (int)reader["legajo"],
-                    reader["nombre"] as string ?? string.Empty,
-                    reader["apellido"] as string ?? string.Empty,
-                    (int)reader["id_tipo_doc"],
-                    reader["num_doc"] as string ?? string.Empty,
-                    reader["fecha_nacimiento"] as DateTime?,
-                    reader["cuil"] as string,
-                    reader["calle"] as string,
-                    reader["altura"] as string,
-                    (int)reader["id_localidad"],
-                    (int)reader["id_genero"],
-                    reader["correo"] as string,
-                    reader["celular"] as string,
-                    (DateTime)reader["fecha_ingreso"]
-                ));
-            }
-            return list;
-        });
-
         public void AddHistorialContrasena(HistorialContrasena historial) => ExecuteNonQuery("sp_historial_contrasena", p =>
         {
             p.AddWithValue("@id_usuario", historial.IdUsuario);
             p.AddWithValue("@contrasena_script", historial.ContrasenaScript);
-        });
-
-        public void AddPersona(Persona persona)
-        {
-            _logger.LogInformation("Agregando persona a la base de datos: Legajo={Legajo}", persona.Legajo);
-            ExecuteNonQuery("sp_insert_persona", p =>
-            {
-                p.AddWithValue("@legajo", persona.Legajo);
-                p.AddWithValue("@nombre", persona.Nombre);
-                p.AddWithValue("@apellido", persona.Apellido);
-                p.AddWithValue("@id_tipo_doc", persona.IdTipoDoc);
-                p.AddWithValue("@num_doc", persona.NumDoc);
-                p.AddWithValue("@fecha_nacimiento", (object?)persona.FechaNacimiento ?? DBNull.Value);
-                p.AddWithValue("@cuil", (object?)persona.Cuil ?? DBNull.Value);
-                p.AddWithValue("@calle", (object?)persona.Calle ?? DBNull.Value);
-                p.AddWithValue("@altura", (object?)persona.Altura ?? DBNull.Value);
-                p.AddWithValue("@id_localidad", persona.IdLocalidad);
-                p.AddWithValue("@id_genero", persona.IdGenero);
-                p.AddWithValue("@correo", (object?)persona.Correo ?? DBNull.Value);
-                p.AddWithValue("@celular", (object?)persona.Celular ?? DBNull.Value);
-            });
-            _logger.LogInformation("Persona con legajo {Legajo} agregada exitosamente.", persona.Legajo);
-        }
-
-        public void AddRespuestaSeguridad(RespuestaSeguridad respuesta) => ExecuteNonQuery("sp_insert_respuesta_seguridad", p =>
-        {
-            p.AddWithValue("@id_usuario", respuesta.IdUsuario);
-            p.AddWithValue("@id_pregunta", respuesta.IdPregunta);
-            p.AddWithValue("@respuesta", respuesta.Respuesta);
         });
 
         public void AddUsuario(Usuario usuario) => ExecuteNonQuery("sp_insert_usuario", p =>
@@ -388,52 +172,9 @@ namespace DataAccess.Repositories
             p.AddWithValue("@CambioContrasenaObligatorio", usuario.CambioContrasenaObligatorio);
         });
 
-        public void UpdatePoliticaSeguridad(PoliticaSeguridad politica) => ExecuteNonQuery("sp_update_politica_seguridad", p =>
-        {
-            p.AddWithValue("@id_politica", politica.IdPolitica);
-            p.AddWithValue("@min_caracteres", (object)politica.MinCaracteres ?? DBNull.Value);
-            p.AddWithValue("@cant_preguntas", (object)politica.CantPreguntas ?? DBNull.Value);
-            p.AddWithValue("@mayus_y_minus", (object)politica.MayusYMinus ?? DBNull.Value);
-            p.AddWithValue("@letras_y_numeros", (object)politica.LetrasYNumeros ?? DBNull.Value);
-            p.AddWithValue("@caracter_especial", (object)politica.CaracterEspecial ?? DBNull.Value);
-            p.AddWithValue("@autenticacion_2fa", (object)politica.Autenticacion2FA ?? DBNull.Value);
-            p.AddWithValue("@no_repetir_anteriores", (object)politica.NoRepetirAnteriores ?? DBNull.Value);
-            p.AddWithValue("@sin_datos_personales", (object)politica.SinDatosPersonales ?? DBNull.Value);
-        });
-
         public void DeleteUsuario(int usuarioId) => ExecuteNonQuery("sp_delete_usuario", p =>
         {
             p.AddWithValue("@id_usuario", usuarioId);
-        });
-
-        public void DeleteRespuestasSeguridadByUsuarioId(int usuarioId) => ExecuteNonQuery(
-            "DELETE FROM respuestas_seguridad WHERE id_usuario = @id_usuario",
-            p => p.AddWithValue("@id_usuario", usuarioId),
-            CommandType.Text
-        );
-
-        public void UpdatePersona(Persona persona) => ExecuteNonQuery("sp_update_persona", p =>
-        {
-            p.AddWithValue("@id_persona", persona.IdPersona);
-            p.AddWithValue("@legajo", persona.Legajo);
-            p.AddWithValue("@nombre", persona.Nombre);
-            p.AddWithValue("@apellido", persona.Apellido);
-            p.AddWithValue("@id_tipo_doc", persona.IdTipoDoc);
-            p.AddWithValue("@num_doc", persona.NumDoc);
-            p.AddWithValue("@fecha_nacimiento", (object?)persona.FechaNacimiento ?? DBNull.Value);
-            p.AddWithValue("@cuil", (object?)persona.Cuil ?? DBNull.Value);
-            p.AddWithValue("@calle", (object?)persona.Calle ?? DBNull.Value);
-            p.AddWithValue("@altura", (object?)persona.Altura ?? DBNull.Value);
-            p.AddWithValue("@id_localidad", persona.IdLocalidad);
-            p.AddWithValue("@id_genero", persona.IdGenero);
-            p.AddWithValue("@correo", (object?)persona.Correo ?? DBNull.Value);
-            p.AddWithValue("@celular", (object?)persona.Celular ?? DBNull.Value);
-            p.AddWithValue("@fecha_ingreso", (object?)persona.FechaIngreso ?? DBNull.Value);
-        });
-
-        public void DeletePersona(int personaId) => ExecuteNonQuery("sp_delete_persona", p =>
-        {
-            p.AddWithValue("@id_persona", personaId);
         });
     }
 }
